correction set_diagonally_dominante_non fini

+++ src/cpp/include/tbsla/cpp/Matrix.hpp
@@ -42,13 +42,13 @@ class Matrix {
     virtual void fill_cqmat(int n_row, int n_col, int c, double q, unsigned int seed_mult = 1, int pr = 0, int pc = 0, int NR = 1, int NC = 1) = 0;
     virtual void fill_random(int n_row, int n_col, double nnz_ratio, unsigned int seed_mult = 1, int pr = 0, int pc = 0, int NR = 1, int NC = 1) = 0;
     virtual void fill_brain(int n_row, int n_col, int* neuron_type, std::vector<std::vector<double> > proba_conn, std::vector<std::unordered_map<int,std::vector<int> > > brain_struct, unsigned int seed_mult, int pr = 0, int pc = 0, int NR = 1, int NC = 1) = 0;
-    virtual void fill_cdistrib(int n_row, int n_col, double nnz, int pr = 0, int pc = 0, int NR = 1, int NC = 1) = 0;
+    virtual void fill_cdistrib(int n_row, int n_col, int nnz, int pr = 0, int pc = 0, int NR = 1, int NC = 1) = 0;
 	
 	virtual void get_row_sums(double* buffer) = 0;
 	virtual void normalize_rows(double* s) = 0;
 	virtual void get_col_sums(double* buffer) = 0;
 	virtual void normalize_cols(double* s) = 0;
-    virtual void set_diag(double* s) = 0;
+	virtual void set_diag(double* s) = 0;
 
   protected:
     int n_row, n_col, f_row, f_col, ln_row, ln_col, pr, pc, NR, NC;
